+++ Panther/include/AST.h
@@ -58,34 +58,16 @@ namespace pcit::panther::AST{
 	};
 
 
-	template<bool IsOptional>
-	class NodeImpl{
+	// access the internal values of this through ASTBuffer
+	//    some funcs are static, but if not get the ASTBuffer from the respective Source
+	class Node{
 		public:
-			NodeImpl(Kind _kind, Token::ID token_id) : kind(_kind), _value{.token_id = token_id} {}
-			NodeImpl(Kind _kind, uint32_t node_index) : kind(_kind), _value{.node_index = node_index} {}
+			constexpr Node(Kind _kind, Token::ID token_id) : kind(_kind), _value{.token_id = token_id} {}
+			constexpr Node(Kind _kind, uint32_t node_index) : kind(_kind), _value{.node_index = node_index} {}
 
-			NodeImpl(const NodeImpl<false>& rhs) requires(IsOptional)
-				: kind(rhs.kind), _value{.node_index = rhs._value.node_index}{}
+			constexpr Node(const Node& rhs) = default;
 
-			NodeImpl(const NodeImpl<IsOptional>&) = default;
-
-
-			NodeImpl() requires(IsOptional) : kind(Kind::None), _value{} {}
-			NodeImpl(std::nullopt_t) requires(IsOptional) : kind(Kind::None), _value{} {}
-
-
-
-			EVO_NODISCARD auto getKind() const -> Kind { return this->kind; }
-
-
-			EVO_NODISCARD auto hasValue() const -> bool {
-				return this->kind != Kind::None;
-			}
-
-			EVO_NODISCARD auto value() const -> NodeImpl<false> requires(IsOptional) {
-				evo::debugAssert(this->hasValue(), "optional node does not have value");
-				return *(NodeImpl<false>*)this;
-			}
+			EVO_NODISCARD constexpr auto getKind() const -> Kind { return this->kind; }
 
 		
 		private:
@@ -98,29 +80,52 @@ namespace pcit::panther::AST{
 				uint32_t node_index; // used by ASTBuffer to get the data
 			} _value;
 
-			static_assert(sizeof(Value) == 4);
-
 
 			friend ASTBuffer;
-			friend NodeImpl<false>;
-			friend NodeImpl<true>;
+			friend std::optional<Node>;
+			friend struct NodeOptInterface;
 	};
 
-	using Node = NodeImpl<false>;
-	using NodeOptional = NodeImpl<true>;
+	struct NodeOptInterface{
+		static constexpr auto init(Node* node) -> void {
+			node->kind = Kind::None;
+		}
+
+		static constexpr auto has_value(const Node& node) -> bool {
+			return node.getKind() != Kind::None;
+		}
+	};
+
 
 	static_assert(sizeof(Node) == 8, "sizeof AST::Node is different than expected");
-	static_assert(sizeof(NodeOptional) == 8, "sizeof AST::NodeOptional is different than expected");
 	static_assert(std::is_trivially_copyable_v<Node>, "AST::Node is not trivially copyable");
-	static_assert(std::is_trivially_copyable_v<NodeOptional>, "AST::NodeOptional is not trivially copyable");
+
+}
 
 
+
+namespace std{
+	
+	template<>
+	class optional<pcit::panther::AST::Node> 
+		: public pcit::core::Optional<pcit::panther::AST::Node, pcit::panther::AST::NodeOptInterface>{
+
+		public:
+			using pcit::core::Optional<pcit::panther::AST::Node, pcit::panther::AST::NodeOptInterface>::Optional;
+			using pcit::core::Optional<pcit::panther::AST::Node, pcit::panther::AST::NodeOptInterface>::operator=;
+	};
+
+}
+
+
+namespace pcit::panther::AST{
+	
 	struct VarDecl{
 		bool isDef;
-		Node ident;
-		NodeOptional type;
+		Token::ID ident;
+		std::optional<Node> type;
 		Node attributeBlock;
-		NodeOptional value;
+		std::optional<Node> value;
 	};
 
 	struct FuncDecl{
@@ -131,19 +136,19 @@ namespace pcit::panther::AST{
 				In,
 			};
 
-			Node ident;
-			NodeOptional type;
+			Node name;
+			std::optional<Node> type; // no type given if name is `this`
 			Kind kind;
 			Node attributeBlock;
 		};
 
 		struct Return{
-			NodeOptional ident;
+			std::optional<Token::ID> ident;
 			Node type;
 		};
 
-		Node ident;
-		NodeOptional templatePack;
+		Node name;
+		std::optional<Node> templatePack;
 		evo::SmallVector<Param> params;
 		Node attributeBlock;
 		evo::SmallVector<Return> returns;
@@ -151,25 +156,25 @@ namespace pcit::panther::AST{
 	};
 
 	struct AliasDecl{
-		Node ident;
+		Token::ID ident;
 		Node type;
 	};
 
 
 	struct Return{
 		Token::ID keyword;
-		NodeOptional label;
-		NodeOptional value;
+		std::optional<Node> label;
+		std::optional<Node> value;
 	};
 
 	struct Block{
-		NodeOptional label;
+		std::optional<Node> label;
 		evo::SmallVector<Node> stmts;
 	};
 
 	struct FuncCall{
 		struct Arg{
-			NodeOptional explicitIdent;
+			std::optional<Node> explicitIdent;
 			Node value;
 		};
 
@@ -179,7 +184,7 @@ namespace pcit::panther::AST{
 
 	struct TemplatePack{
 		struct Param{
-			Node ident;
+			Token::ID ident;
 			Node type;
 		};
 
@@ -227,8 +232,8 @@ namespace pcit::panther::AST{
 
 	struct AttributeBlock{
 		struct Attribute{
-			Node name;
-			NodeOptional arg;
+			Token::ID attribute;
+			std::optional<Node> arg;
 		};
 
 		evo::SmallVector<Attribute> attributes;