+++ Panther/include/diagnostics.h
@@ -47,6 +47,8 @@ namespace pcit::panther{
 		SemaEncounteredKindNone,            // S1
 		SemaInvalidGlobalStmt,              // S2
 		SemaAlreadyDefined,                 // S3
+		SemaNamedReturnParamIsTypeVoid,     // S4,
+		SemaVoidWithQualifiers,             // S5,
 
 		MiscFileDoesNotExist,               // M1
 		MiscLoadFileFailed,                 // M2
@@ -92,6 +94,8 @@ namespace pcit::panther{
 			break; case DiagnosticCode::SemaEncounteredKindNone:            return "S1";
 			break; case DiagnosticCode::SemaInvalidGlobalStmt:              return "S2";
 			break; case DiagnosticCode::SemaAlreadyDefined:                 return "S3";
+			break; case DiagnosticCode::SemaNamedReturnParamIsTypeVoid:     return "S4";
+			break; case DiagnosticCode::SemaVoidWithQualifiers:             return "S5";
 
 			break; case DiagnosticCode::MiscFileDoesNotExist:               return "M1";
 			break; case DiagnosticCode::MiscLoadFileFailed:                 return "M2";