+++ Panther/src/TypeManager.cpp
@@ -17,8 +17,8 @@ namespace pcit::panther{
 			delete builtin;
 		}
 
-		for(Type* type : this->types){
-			delete type;
+		for(TypeInfo* type_info : this->types){
+			delete type_info;
 		}
 	}
 	
@@ -41,6 +41,7 @@ namespace pcit::panther{
 		this->builtins.emplace_back(new BaseType::Builtin(Token::Kind::TypeBool));
 		this->builtins.emplace_back(new BaseType::Builtin(Token::Kind::TypeChar));
 		this->builtins.emplace_back(new BaseType::Builtin(Token::Kind::TypeRawPtr));
+
 		this->builtins.emplace_back(new BaseType::Builtin(Token::Kind::TypeCShort));
 		this->builtins.emplace_back(new BaseType::Builtin(Token::Kind::TypeCUShort));
 		this->builtins.emplace_back(new BaseType::Builtin(Token::Kind::TypeCInt));
@@ -67,14 +68,58 @@ namespace pcit::panther{
 	}
 
 
+	//////////////////////////////////////////////////////////////////////
+	// type
 
-
-	auto TypeManager::getType(Type::ID id) const -> const Type& {
+	auto TypeManager::getTypeInfo(TypeInfo::ID id) const -> const TypeInfo& {
 		const auto lock = std::shared_lock(this->types_mutex);
 		return *this->types[id.get()];
 	}
 
+	auto TypeManager::getOrCreateTypeInfo(TypeInfo&& lookup_type_info) -> TypeInfo::ID {
+		const auto lock = std::unique_lock(this->types_mutex);
 
+		for(uint32_t i = 0; const TypeInfo* type_info : this->types){
+			if(*type_info == lookup_type_info){
+				return TypeInfo::ID(i);
+			}
+		
+			i += 1;
+		}
+
+		const auto new_type_id = TypeInfo::ID(uint32_t(this->types.size()));
+		this->types.emplace_back(new TypeInfo(std::forward<TypeInfo>(lookup_type_info)));
+		return new_type_id;
+	}
+
+
+	//////////////////////////////////////////////////////////////////////
+	// function
+
+	auto TypeManager::getFunction(BaseType::Function::ID id) const -> const BaseType::Function& {
+		const auto lock = std::shared_lock(this->functions_mutex);
+		return *this->functions[id.get()];
+	}
+
+	auto TypeManager::getOrCreateFunction(BaseType::Function lookup_func) -> BaseType::ID {
+		const auto lock = std::unique_lock(this->functions_mutex);
+		
+		for(uint32_t i = 0; const BaseType::Function* function : this->functions){
+			if(*function == lookup_func){
+				return BaseType::ID(BaseType::Kind::Function, i);
+			}
+		
+			i += 1;
+		}
+
+		const auto new_id = BaseType::ID(BaseType::Kind::Function, uint32_t(this->functions.size()));
+		this->functions.emplace_back(new BaseType::Function(lookup_func));
+		return new_id;
+	}
+
+
+	//////////////////////////////////////////////////////////////////////
+	// builtin
 
 	auto TypeManager::getBuiltin(BaseType::Builtin::ID id) const -> const BaseType::Builtin& {
 		const auto lock = std::shared_lock(this->builtins_mutex);
@@ -90,20 +135,15 @@ namespace pcit::panther{
 	}
 
 	auto TypeManager::get_or_create_builtin_base_type_impl(const BaseType::Builtin& lookup_type) -> BaseType::ID {
-		{
-			const auto lock = std::shared_lock(this->builtins_mutex);
-
-			// TODO: improve lookup times with maps?
-			for(uint32_t i = 0; const BaseType::Builtin* builtin_type : this->builtins){
-				if(*builtin_type == lookup_type){
-					return BaseType::ID(BaseType::Kind::Builtin, i);
-				}
-			
-				i += 1;
-			}
-		}
-
 		const auto lock = std::unique_lock(this->builtins_mutex);
+
+		for(uint32_t i = 0; const BaseType::Builtin* builtin_type : this->builtins){
+			if(*builtin_type == lookup_type){
+				return BaseType::ID(BaseType::Kind::Builtin, i);
+			}
+		
+			i += 1;
+		}
 
 		const auto new_id = BaseType::ID(BaseType::Kind::Builtin, uint32_t(this->builtins.size()));
 		this->builtins.emplace_back(new BaseType::Builtin(lookup_type));